+++ src/parser/block/blockquote.ts
@@ -5,6 +5,7 @@ import { autolink } from '../autolink';
 import { parse } from '../api/parse';
 import { defrag, suppress } from '../util';
 import { figure } from '../../util/figure';
+import { footnote } from '../../util/footnote';
 import { html } from 'typed-dom';
 
 export const segment: BlockquoteParser.SegmentParser = block(union([
@@ -45,10 +46,12 @@ const source: BlockquoteParser.SourceParser = lazy(() => fmap(
       convert(unindent, source)),
     rewrite(
       some(contentline, opener),
-      convert(unindent, source => [[suppress(figure_(parse(source)))], ''])),
+      convert(unindent, source => [[suppress(render(parse(source)))], ''])),
   ])),
   ns => [html('blockquote', ns)]));
 
-function figure_(source: DocumentFragment): DocumentFragment {
-  return void figure(source), source;
+function render(source: DocumentFragment): DocumentFragment {
+  void figure(source);
+  void footnote(source, { annotation: html('ol'), authority: html('ol') });
+  return source;
 }